Inline more into profile & delete unused components
